requirements.txt
falcon==2.0.0
gunicorn==19.9.0
meinheld==1.0.1
python-rapidjson==0.7.1
setproctitle==1.1.10
vyper-config==0.4.0
redis==3.2.1
psycopg2==2.8.2
SQLAlchemy==1.3.3
